Example commit in the bootcamp. I did NOT finish debugging, and I'm not confident about the debugging I did do.
