+++ c.py
@@ -1,296 +1,3 @@
-def manova_eta_fdr(data_tbl: pd.DataFrame, dep_vars: [str], cat_vars: [str]):
-    """
-    Do a multi-test, adjust stuff, and get some squared number at the end.
-
-    Parameters:
-    data (pd.DataFrame): The input data frame.
-    dep_vars (list of str): The names of the dependent variables.
-    cat_vars (list of str): The names of the categorical variables.
-
-    Returns:
-    tuple: MANOVA results, ANOVA test results with eta squared, pairwise comparison results with FDR adjusted p-values.
-    """
-    categorical_frm = ' + '.join(cat_vars)
-    frm = f'{"+".join(dep_vars)} ~ {categorical_frm}'
-    manova = MANOVA.from_frm(frm, data=data_tbl)
-
-    stat_res = {}
-    pairwise_results = {}
-    eta_num_results = {}
-
-    for dependent_var in dep_vars:
-        # Define the ANOVA frm for each dependent variable
-        anova_frm = f'{dependent_var} ~ {categorical_frm}'
-
-        # Fit the ANOVA model
-        model = ols(anova_frm, data=data_tbl).fit()
-
-        # Perform ANOVA
-        anova_table = sm.stats.anova_lm(model, typ=2)
-
-        # Calculate eta squared
-        ss_effect = anova_table['sum_sq'].iloc[0]
-        ss_total = anova_table['sum_sq'].sum()
-        eta_num = ss_effect / ss_total
-
-        # Add eta squared to the ANOVA table
-        anova_table['eta_sq'] = [eta_num] + [None] * (len(anova_table) - 1)
-
-        # Store ANOVA table
-        stat_res[dependent_var] = anova_table
-
-        # Perform pairwise comparisons using Tukey's HSD test
-        tukey_result = pairwise_tukeyhsd(endog=data_tbl[dependent_var], groups=data_tbl[cat_vars[0]], alpha=0.05)
-
-        # Extract p-values from the pairwise comparison results
-        p_values = tukey_result.pvalues
-
-        # Perform FDR adjustment on the p-values
-        _, pvals_corrected, _, _ = multipletests(p_values, alpha=0.05, method='fdr_bh')
-
-        tukey_result_data = np.array(tukey_result._results_table.data)
-        # Create a DataFrame for the pairwise comparison results with FDR adjusted p-values
-        pairwise_result = pd.DataFrame(data={
-            'group1': tukey_result_data[1:, 0],
-            'group2': tukey_result_data[1:, 1],
-            'meandiff': tukey_result_data[1:, 2],
-            'p-adj': tukey_result_data[1:, 3],
-            'lower': tukey_result_data[1:, 4],
-            'upper': tukey_result_data[1:, 5],
-            'reject': tukey_result_data[1:, 6],
-            'pvals_corrected': pvals_corrected
-        })
-
-        # Store pairwise results
-        pairwise_results[dependent_var] = pairwise_result
-
-    return manova, stat_res, pairwise_results
-
-
-def lasso_regr(data_tbl: pd.DataFrame, in_features: [str], out_features: [str]):
-    input = data_tbl[in_features]
-    output = data_tbl[out_features]
-    model = MultiTaskLassoCV().fit(input, output)
-    return model
-
-
-def homo_check(data_tbl: pd.DataFrame, y: pd.Series, alpha=0.05, with_conclusion_print=False) -> Tuple[
-    bool, float, float, float, float]:
-    """
-    See if the spread of data is even by running a specific test on it.
-
-    This function performs the Breusch-Pagan test for heteroscedasticity. It uses different
-    criteria based on the sample size to determine homoscedasticity.
-
-    Args:
-        data_tbl (pd.DataFrame): Feature matrix.
-        y (pd.Series): Target variable.
-        alpha (float, optional): The significance level for the test. Defaults to 0.05.
-        with_conclusion_print (bool): print the conclusion of the test.
-    Returns:
-        Tuple[bool, float, float, float, float]: A tuple containing:
-            - bool: True if variance is likely homoscedastic, False otherwise.
-              For sample sizes <= 30, only the F-test p-value is considered.
-              For larger samples, both LM and F-test p-values must exceed alpha.
-            - float: Lagrange Multiplier (LM) statistic
-            - float: p-value for the LM statistic
-            - float: F-value
-            - float: p-value for the F-statistic
-
-    Notes:
-        - For small samples (n <= 30), only the F-test is used due to its better small-sample properties.
-        - For larger samples, both tests must indicate homoscedasticity for the function to return True.
-
-    Reference:
-    Breusch, T. S., & Pagan, A. R. (1979). A simple test for heteroscedasticity and
-    random coefficient variation. Econometrica, 47(5), 1287-1294.
-    https://www.jstor.org/stable/1911963
-    """
-    input_with_const = sm.add_constant(data_tbl)
-
-    # Fit the model
-    model = sm.OLS(y, input_with_const).fit()
-    lm, lm_pvalue, fvalue, f_pvalue = het_breuschpagan(model.resid, model.model.exog)
-    if len(data_tbl) <= 30:
-        is_homoscedasticity = lm_pvalue > alpha and f_pvalue > alpha
-
-    else:
-        is_homoscedasticity = lm_pvalue > alpha and f_pvalue > alpha
-    if with_conclusion_print:
-        print_homoscedasticity_conclusion(is_homoscedasticity, lm_pvalue, f_pvalue, alpha, len(data_tbl))
-    return is_homoscedasticity, lm, lm_pvalue, fvalue, f_pvalue
-
-
-def auto_corr_check(leftovers: np.ndarray, alpha=0.05, with_conclusion_print=False) -> Tuple[bool, float, float]:
-    """
-    Check for autocorrelation using the Durbin-Watson test and Ljung-Box test.
-
-    Args:
-        leftovers (np.ndarray): The leftovers from a linear regression model.
-        alpha (float): The significance level for the test.
-        with_conclusion_print (bool): print the conclusion of the test.
-
-    Returns:
-        Tuple[bool, float, float]: A tuple containing:
-            - bool: True if no significant autocorrelation (Ljung-Box p-value > alpha), False otherwise.
-            - float: The p-value of the Ljung-Box test.
-            - float: The Durbin-Watson statistic.
-
-    References:
-    Durbin, J., & Watson, G. S. (1951). Testing for serial correlation in least
-    squares regression. II. Biometrika, 38(1/2), 159-177.
-    https://www.jstor.org/stable/2332325
-
-    Ljung, G. M., & Box, G. E. P. (1978). On a measure of lack of fit in time series models.
-    Biometrika, 65(2), 297-303.
-    https://doi.org/10.1093/biomet/65.2.297
-    :param with_conclusion_print:
-    """
-    dw_statistic = durbin_watson(leftovers)
-    lb_test = acorr_ljungbox(leftovers, lags=[1], return_data_tbl=True)
-    lb_p_value = lb_test['lb_pvalue'].values[0]
-    if with_conclusion_print:
-        print_autocorrelation_conclusion(lb_p_value > alpha, lb_p_value, dw_statistic, alpha)
-    return lb_p_value > alpha, lb_p_value, dw_statistic
-
-
-def linear_test_res(is_linear: bool, alpha: float):
-    """
-    Print the conclusion from the linearity test.
-    """
-    if is_linear:
-        print("Conclusion: The relationship appears to be linear.")
-    else:
-        print(f"Conclusion: The relationship may not be linear (p-value <= {alpha}).")
-
-
-def desc_stats(data_tbl: pd.DataFrame, columns: [str], with_print=False):
-    desc_data_tbl = data_tbl.copy()
-    data = [('var', 'count', 'mean', 'std', 'min', '25%', '50%', '75%', 'max', 'ci lower', 'ci upper')]
-    for col in columns:
-        col_data = desc_data_tbl[~desc_data_tbl[col].isna()][col].astype(int)
-        desc = col_data.describe()
-        sums = col_data.tolist()
-        mean = np.mean(sums)
-        s = sem(sums)
-        ci = t.interval(0.95, len(sums) - 1, loc=mean, scale=s)
-        data.append((col, desc['count'], mean, np.std(sums), desc['min'], desc['25%'], desc['50%'], desc['75%'],
-                     desc['max'], ci[0], ci[1]))
-
-    if with_print:
-        print(tabulate(data[1:], headers=data[0], tablefmt='fancy_grid'))
-    return data
-
-
-def cat_count_plot(data_tbl, category_column, subplot_column=None, title="", x_label=""):
-    """
-    Counts the number of items in each category of the specified column in the DataFrame
-    and plots the counts as a bar chart. If subplot_column is provided, creates a subplot
-    for each unique value in the subplot_column.
-
-    Parameters:
-    data_tbl (pd.DataFrame): The DataFrame containing the data.
-    category_column (str): The name of the category column to count items in.
-    subplot_column (str, optional): The name of the column to create subplots for. Default is None.
-
-    Returns:
-    dict: A dictionary with the counts of each category for each subplot (if subplot_column is provided).
-    """
-    plot_data_tbl = data_tbl.copy()
-    if category_column not in plot_data_tbl.columns:
-        raise ValueError(f"Column '{category_column}' does not exist in the DataFrame.")
-
-    if subplot_column and subplot_column not in plot_data_tbl.columns:
-        raise ValueError(f"Column '{subplot_column}' does not exist in the DataFrame.")
-    palette = palettes[category_column] if category_column in palettes else None
-    order = orders[category_column] if category_column in orders else None
-
-    if subplot_column:
-        unique_values = plot_data_tbl[subplot_column].unique()
-        num_subplots = len(unique_values)
-        fig, axes = plt.subplots(1, num_subplots, figsize=(8 * num_subplots, 6))
-        counts_dict = {}
-
-        for i, value in enumerate(unique_values):
-            ax = axes[i] if num_subplots > 1 else axes
-            ax.set_facecolor("white")
-            subset = plot_data_tbl[plot_data_tbl[subplot_column] == value]
-            counts = subset[category_column].value_counts()
-            if order:
-                index_order = sorted(counts.index.to_list(), key=lambda x: order[x])
-                counts = counts.reindex(index_order)
-            colors = ['#FF5733'] * len(counts)
-            if palette is not None:
-                for j, col in enumerate(counts.to_dict().keys()):
-                    colors[j] = palette[col]
-            total = counts.sum()
-            counts.plot(kind='bar', ax=ax, color=colors)
-            for p in ax.patches:
-                height = p.get_height()
-                percentage = 100 * height / total
-                ax.text(
-                    p.get_x() + p.get_width() / 2,
-                    height,
-                    f'{percentage:.1f}%\nN:{height}',
-                    ha='center',
-                    va='bottom'
-                )
-            ax.set_suptitle(title)
-            if x_label:
-                ax.set_xlabel(x_label)
-            else:
-                ax.set_xlabel(category_column)
-            ax.set_ylabel('')
-            ax.set_xticks(ax.get_xticks())
-            ax.set_xticklabels(ax.get_xticklabels(), rotation=30, ha='center')
-            ax.set_yticks([])
-            counts_dict[value] = counts
-            ax.spines['right'].set_visible(False)
-            ax.spines['left'].set_visible(False)
-            ax.spines['top'].set_visible(False)
-        plt.tight_layout()
-        plt.show()
-        return counts_dict
-
-    else:
-        category_counts = plot_data_tbl[category_column].value_counts()
-        if order:
-            index_order = sorted(category_counts.index.to_list(), key=lambda x: order[x])
-            category_counts = category_counts.reindex(index_order)
-        plt.figure(figsize=(10, 6))
-        colors = ['#FF5733'] * len(category_counts)
-        if palette is not None:
-            for i, col in enumerate(category_counts.to_dict().keys()):
-                colors[i] = palette[col]
-        category_counts.plot(kind='bar', color=colors)
-        total = category_counts.sum()
-        ax = plt.gca()
-        ax.set_facecolor("white")
-        for p in plt.gca().patches:
-            height = p.get_height()
-            percentage = 100 * height / total
-            ax.text(
-                p.get_x() + p.get_width() / 2,
-                height,
-                f'{percentage:.1f}%\nN:{height}',
-                ha='center',
-                va='bottom'
-            )
-        plt.suptitle(title)
-        if x_label:
-            ax.set_xlabel(x_label)
-        else:
-            ax.set_xlabel(category_column)
-        plt.ylabel('')
-        ax.spines['right'].set_visible(False)
-        ax.spines['left'].set_visible(False)
-        ax.spines['top'].set_visible(False)
-        plt.xticks(rotation=30, ha='center')
-        plt.yticks([])
-        plt.show()
-
-        return category_counts
-
 
 def interact_plot(data_tbl: pd.DataFrame, first_cat: str, second_cat: str, depended_vars: list,
                                        aggregation_function):
